Deal with int overflow for EGB files.
git-svn-id: http://encog-cs.googlecode.com/svn/trunk/encog-core@681 0d3f3ea6-ac51-0410-a1de-edb5cd2a854a

+++ encog-core-cs/ML/Data/Buffer/EncogEGBFile.cs
@@ -277,9 +277,9 @@ namespace Encog.ML.Data.Buffer
         /// </summary>
         /// <param name="row">The row to calculate for.</param>
         /// <returns>The index.</returns>
-        private int CalculateIndex(int row)
+        private long CalculateIndex(long row)
         {
-            return HeaderSize + (row*_recordSize);
+            return (long)HeaderSize + (row*(long)_recordSize);
         }
 
         /// <summary>
